Update PackageId and include README in NuGet packages
Updated `PipeR.AspNetCore.csproj` to change the `PackageId`
to `PipeR.Extensions.AspNetCore`. Added `README.md` as the
`PackageReadmeFile` and ensured it is included in the NuGet
package.

Updated `PipeR.Core.csproj` to add `README.md` as the
`PackageReadmeFile` and ensured it is included in the NuGet
package.

Both project files now include the `README.md` file in their
respective NuGet packages to provide additional documentation
for package consumers.